remove duplicated test
